python-legacy/packages/bss-telemetry/pyproject.toml
[project]
name = "bss-telemetry"
version = "0.1.0"
requires-python = ">=3.12"
description = "OpenTelemetry bootstrap + BSS semantic conventions for cross-service tracing."
dependencies = [
    "structlog>=24.0",
    "pydantic-settings>=2.0",
    # Targets of the auto-instrumentors — declared so configure_telemetry()
    # can wire all four cleanly in any service. fastapi/httpx are already
    # transitively present via every service; asyncpg/aio-pika are explicit
    # in write/MQ services. Listing them here makes the bootstrap honest.
    "fastapi>=0.110",
    "httpx>=0.27",
    "asyncpg>=0.29",
    "aio-pika>=9.0",
    "opentelemetry-api==1.39.0",
    "opentelemetry-sdk==1.39.0",
    "opentelemetry-exporter-otlp-proto-http==1.39.0",
    "opentelemetry-instrumentation-fastapi==0.60b0",
    "opentelemetry-instrumentation-httpx==0.60b0",
    "opentelemetry-instrumentation-asyncpg==0.60b0",
    "opentelemetry-instrumentation-aio-pika==0.60b0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
